backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
cloudinary==1.36.0
python-dotenv==1.0.0
torch==2.1.1
torchvision==0.16.1
tensorflow-cpu==2.15.0
numpy==1.26.2
Pillow==10.1.0
pandas==2.1.3
joblib==1.3.2
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78
xgboost==2.0.2
pydantic==2.5.2
python-multipart==0.0.6
aiofiles==23.2.1
